split FilterHandler in Filter and handler. Multiple handlers for one filter are now possible. Bump version of common-utils
