+++ composer.json
@@ -6,7 +6,7 @@
     }
   },
   "require": {
-    "bendani/php-common/utils": "1.0.*",
+    "bendani/php-common/utils": "2.0.*",
     "bendani/php-common/uiframework": "1.3.*"
   },
   "repositories": [
