+++ src/main/model/FilterHandler.php
@@ -5,15 +5,6 @@ namespace Bendani\PhpCommon\FilterService\Model;
 interface FilterHandler
 {
 
-    public function handleFilter(Filter $filter);
+    public function handleFilter(FilterValue $filter, $object = null);
 
-    public function getFilterId();
-
-    public function getGroup();
-
-    public function getType();
-
-    public function getField();
-
-    public function getSupportedOperators();
 }
