+++ src/main/model/FilterManager.php
@@ -2,25 +2,39 @@
 
 namespace Bendani\PhpCommon\FilterService\Model;
 
+use Bendani\PhpCommon\Utils\Ensure;
+
 class FilterManager
 {
 
     /** @var array  */
     private $filters;
+    /** @var array  */
+    private $handlers = array();
+
     private $filtersAsList;
 
     public function __construct($filters)
     {
         $this->filtersAsList = $filters;
+        /** @var Filter $filter */
         foreach($filters as $filter){
             $this->filters[$filter->getFilterId()] = $filter;
         }
     }
 
-    public function handle(Filter $filter){
+    public function handle($handlerGroupId, FilterValue $filter, $object = null){
+        Ensure::objectNotNull('handlerGroup', $this->handlers[$handlerGroupId]);
+
         /** @var FilterHandler $handler */
-        $handler = $this->filters[$filter->getId()];
-        return $handler->handleFilter($filter);
+        $handler = $this->$this->handlers[$handlerGroupId][$filter->getId()];
+        Ensure::objectNotNull('handler for filter', $this->handlers[$handlerGroupId]);
+
+        return $handler->handleFilter($filter, $object);
+    }
+
+    public function registerHandlers($handlerGroupId, $handlers){
+        $handlers[$handlerGroupId] = $handlers;
     }
 
     public function getFilters()
