+++ src/main/model/FilterValue.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace Bendani\PhpCommon\FilterService\Model;
+
+
+interface FilterValue
+{
+
+    function getId();
+
+    function getValue();
+
+    function getOperator();
+
+}
